+++ index.js
@@ -5,10 +5,16 @@ const {idForBlock} = require('./hashes');
 const {idForTransactionComponents} = require('./hashes');
 const {noLocktimeIdForTransaction} = require('./transactions');
 const {numberAsCompactInt} = require('./numbers');
+const {p2pkhOutputScript} = require('./script');
+const {p2shOutputScript} = require('./script');
+const {p2trOutputScript} = require('./script');
+const {p2wpkhOutputScript} = require('./script');
+const {p2wshOutputScript} = require('./script');
 const {previousBlockId} = require('./hashes');
 const {queryTransactions} = require('./transactions');
 const {scriptAsScriptElements} = require('./script');
 const {scriptElementsAsScript} = require('./script');
+const {unsignedTxFromPsbt} = require('./transactions');
 
 module.exports = {
   compactIntAsNumber,
@@ -18,8 +24,14 @@ module.exports = {
   idForTransactionComponents,
   noLocktimeIdForTransaction,
   numberAsCompactInt,
+  p2pkhOutputScript,
+  p2shOutputScript,
+  p2trOutputScript,
+  p2wpkhOutputScript,
+  p2wshOutputScript,
   previousBlockId,
   queryTransactions,
   scriptAsScriptElements,
   scriptElementsAsScript,
+  unsignedTxFromPsbt,
 };