+++ transactions/index.js
@@ -2,10 +2,12 @@ const componentsOfTransaction = require('./components_of_transaction');
 const noLocktimeIdForTransaction = require('./no_locktime_id_for_transaction');
 const parseTransaction = require('./parse_transaction');
 const queryTransactions = require('./query_transactions');
+const unsignedTxFromPsbt = require('./unsigned_tx_from_psbt');
 
 module.exports = {
   componentsOfTransaction,
   noLocktimeIdForTransaction,
   parseTransaction,
   queryTransactions,
+  unsignedTxFromPsbt,
 };